requirements.txt
Django==2.2.3
channels
channels_redis
python_dotenv
